archive/experimental_implementations/requirements.txt
# Super-Digimon GraphRAG System Requirements

# Core Database Drivers
neo4j==5.18.0
sqlalchemy==2.0.25
faiss-cpu==1.7.4

# MCP Server Framework
mcp==0.9.1
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP and Processing
spacy==3.7.2
sentence-transformers==2.3.1
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Graph Analysis
networkx==3.2.1

# Document Processing
pypdf==3.17.4
PyMuPDF==1.23.8
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# API Support
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Development Tools
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
types-requests==2.31.0.10

# Testing with Real Databases
testcontainers[neo4j]==3.7.1

# Caching (optional)
redis==5.0.1

# LLM Integration
openai>=1.0.0